app/requirements.txt
langchain
langchain-ollama
fastapi
uvicorn
langchain-community
faiss-cpu
requests
PyYAML
httpx
pytest